requirements.txt
zhipuai>=2.0.0
numpy>=1.24.0
openai>=1.0.0
pyyaml>=6.0
anthropic>=0.18.0
